+++ requirements_dev.txt
@@ -1,7 +1,8 @@
 black>=22.1.0
+build>=1.2.1
 coverage>=6.3.2
 flake8>=4.0.1
 ipdb>=0.13.9
 mypy>=0.950
-pytype>=2022.4.26
 pytest>=7.1.2
+pytype>=2022.4.26
