+++ setup.cfg
@@ -1,0 +1,43 @@
+[metadata]
+name = caskdb
+version = 0.1.0
+description = Disk based Log Structured Hash Table Store
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/avinassh/py-caskdb
+author = Avinash Sajjanshetty
+author_email = opensource@avi.im
+license = MIT
+license_file = LICENSE
+classifiers =
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: 3.8
+    Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
+    Programming Language :: Python :: 3.12
+    Programming Language :: Python :: Implementation :: CPython
+    Typing :: Typed
+
+[options]
+packages = find:
+python_requires = >=3.8
+
+[options.extras_require]
+dev =
+    black>=22.1.0
+    build>=1.2.1
+    coverage>=6.3.2
+    flake8>=4.0.1
+    ipdb>=0.13.9
+    mypy>=0.950
+    pytest>=7.1.2
+    pytype>=2022.4.26
+
+
+[options.package_data]
+caskdb =
+    example.py
