+++ tests/test_disk_store.py
@@ -3,7 +3,7 @@ import tempfile
 import typing
 import unittest
 
-from disk_store import DiskStorage
+from caskdb.disk_store import DiskStorage
 
 
 class TempStorageFile:
